requirements.txt
absl-py==0.9.0
alipay==0.7.4
astor==0.8.1
backports.csv==1.0.7
captcha==0.3

chardet==3.0.4
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0.2
django-cors-headers==3.2.1
django-crispy-forms==1.7.2
django-filter==2.2.0
django-formtools==2.1
django-import-export==1.2.0
django-redis==4.10.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.17.1
gast==0.3.2
grpcio==1.26.0
h5py==2.10.0
httplib2==0.13.1
idna==2.8
jdcal==1.4.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.16.4
odfpy==1.4.0
openpyxl==2.6.3
Pillow==6.1.0
protobuf==3.11.2
pycryptodomex==3.7.2
PyExecJS==1.5.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.6
python-alipay-sdk==1.10.1
python-dateutil==2.8.1
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
retrying==1.3.3
scipy==1.4.1
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
urllib3==1.25.6
Werkzeug==0.16.0
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
